Campus Heat Map (MQTT, REST, Android, Beacons)
•	Created an Android app which listens to specific beacons and reports
the user to the node server.
•	Node server which receives the beacons, simulates the number of users
and maintains count of users in 4 different areas. The server
communicates with the boards using pub/sub in MQTT.
•	The server also offers a REST API to android and the mbed boards use
LWM2M objects and REST to flash the light on the board.
•	The mbed boards were also used to simulate a user, which on button
press of an mbed board in an area, increased count of users in that
area and when the number of users in that area are more than a
threshold, then a light is flashed on the board.